+++ src/ConquerSpace/gui/game/planetdisplayer/PlanetIndustry.java
@@ -279,9 +279,9 @@ public class PlanetIndustry extends JPanel {
 
             String inputString = "Input: ";
 
-            for (Map.Entry<Good, Integer> entry : factory.getProcess().input.entrySet()) {
+            for (Map.Entry<Good, Double> entry : factory.getProcess().input.entrySet()) {
                 Good key = entry.getKey();
-                Integer val = entry.getValue();
+                Double val = entry.getValue();
                 inputString = inputString + key.getName();
                 inputString = inputString + " amount " + val;
                 inputString = inputString + ", ";
@@ -290,9 +290,9 @@ public class PlanetIndustry extends JPanel {
             JLabel input = new JLabel(inputString);
 
             String outputString = "Output: ";
-            for (Map.Entry<Good, Integer> entry : factory.getProcess().output.entrySet()) {
+            for (Map.Entry<Good, Double> entry : factory.getProcess().output.entrySet()) {
                 Good key = entry.getKey();
-                Integer val = entry.getValue();
+                Double val = entry.getValue();
                 outputString = outputString + key.getName();
                 outputString = outputString + " amount " + val;
                 outputString = outputString + ", ";
